fix(kleat): A few small changes
Don't set defaultMode as the default is fine here (it appears
that Halyard never set it).

Increase the memory usage since we're in a container and there
isn't anything else running. (We don't set this to 100% as it's
only for the size of the heap, and we want to leave some for the
JVM's non-heap memory.)