feat(config): Add base spinnaker.yml
When deploying to kubernetes, the spinnaker.yml should in general
always be the same as it's just pointing to the Kubernetes service
in front of each deployment. Based on this, we'll include it in the
base kustomization and add it to all of the service configs.

There are still some questions around how to best override this,
particularly to specify an external redis. I actually think that
we may want to suggest that users map the 'redis.spinnaker' to
an Endpoint object pointing outside the cluster, but that is something
we can experiment with later.

For now, I'm only adding entries relevant to the services we're supporting;
we'll add things in as we need them.