feat(install): Add namespace and tag
Let's add the namespace spinnaker to the base kustomization, as
that's probably more reasonable than the default namespace. This
can be overriden in an overlay, but let's avoid the need to override
if you're fine with spinnaker.

Also, add the spinnaker-master-latest-unvalidated tag to the clouddriver
deployment. When it was absent the deploy would fail unless you specified
an override because we don't tag any images with :latest. Let's at least
keep the base deployment valid even though this will almost always be
overriden.

In principle at some point it might be nice to have a :spinnaker-stable
tag pointing at the latest patch release that we could default to here,
but for now let's just use master-latest-unvalidated.