refactor(core): Move all labels to the kustomization.yml file
In some places we're including labels directly in the kubernetes
manifests, and in others we're adding these in the kustomization.yml.

I propose to standardize on this:
* The individual manifests won't have labels (or other label fields
that kustomize can add)
* The kustomizeation.yml alongside each service/deployment adds
the necessary labels (generally the instance and name based on
the service)
* In the case of services that are always overridden by an HA version
(ex: clouddriver) I still added raw 'clouddriver' labels in the base
kustomization just so it would be valid on its own even if we are
always adding different labels.

The advantage is less duplication of labels (and reducing copy/paste
mistakes of forgetting to update one label). The disadvantage is that
it's a bit less clear exactly what labels get generated (ie, we no longer
explicitly include a selector, etc.)

To test this, I generated a full deployment manifest by running
kustomize before and after, and found no differences. (Or rather I
did find differences, but then I fixed them so that by the time you
read this commit there are no differences.)