feat(config): Add fiat, gate, and igor configs
There is nothing special that had to be considered for these
services as far as HA; just adding the same YAML as we had
for other non-HA services.

(At some point we can decide whether to further factor things out
to reduce copy-pase, but for now I think our initial decision to
favor duplication over overly-complex factoring is still the
correct one.)

One complication for Gate is that the health check depends on
whether it has HTTPS enabled, which depends on the halconfig.
For now I've just made the check HTTPS enabled but we'll need
to think about how to support both.