fix(core): Fix service-discovery config (#13)
I'm getting an error when trying to use the base as a github
link, which I believe is because we are trying to access a
resource in a sibling directory. It looks like this may be a
bug in kustomize, but I think we can just move this to the
core/ directory as a workaround.