feat(security): change UID/GID to 10111
(Breaking Change) All Spinnaker microservices (with the exception of Halyard and Deck) now have explicit GIDs/UIDs in their Dockerfile.

If you are using a version of the kustomization-base that includes this commit with a pre-1.22 version of Spinnaker, you will need to include a patch that sets the GID/UID to 1000 (or whatever you've set the GID/UID to in a custom Dockerfile).