fix(security): remove Deck security context block
I should have removed this with the commit that changed all non-Deck services to use the new Spinnaker GID/UID ([PR](https://github.com/spinnaker/kustomization-base/pull/30)).

Plumpy determined that it would be non-trivial to update Deck's Dockerfile to use the new GID/UID and we decided to leave as-is for now. Any user building a custom Deck image can add a custom security context as a patch.